+++ packages/contracts/test/utils/combinatorial_utils.ts
